fix(merge): resolve 1.0.0-rc1 compile fallout after upstream merge
Post-merge Windows compile pass for tools/shell/pager/bin:

- workspace: wrap confine PolicyDeny in PermissionResolution
- shell: adopt upstream workflow host concurrency API; restore
  cancel_pending_recap_for_new_prompt; CancelOptions tuple form;
  SubagentsConfig Default fields; SubagentRequest.allowed_paths;
  CoordinatorConfig.max_concurrent alignment; upstream acp_agent API
- pager-render: restore accent_feedback theme channel; format_bytes u64
- pager: walkdir dep for fs_size; MaterializeCtx restore flags; Cow
  labels/placeholders for Shortcut/PromptStyle

+++ crates/codegen/xai-grok-pager-render/src/render/image_overlay.rs
@@ -249,7 +249,7 @@ fn render_image_overlay_inner(
             }
             lines.push(Line::from(format!(
                 "Size: {}",
-                format_bytes(image.byte_len)
+                format_bytes(image.byte_len as u64)
             )));
             lines.push(Line::from("Preview unavailable"));
             let paragraph = Paragraph::new(lines)
